mingw: don't fetch mesa_wgl.h + update summary blurb
mesa_wgl.h can no longer be fetched from upstream and since it's no
longer used anyway we don't fetch this any more. This also updates
the blurb printed after fetching dependencies to show how to run
./configure so we pass --enable-wgl not --enable-stub-winsys and
to also pass the -I path for the cogl-cross/include directory which has
the latest gl.h we fetched so the build doesn't try and use the headers
shipped with the mingw toolchain which may be out-of-date.

Reviewed-by: Neil Roberts <neil@linux.intel.com>