tests: avoid redundant setting of projection/modelview
This removes some redundant initializing of the modelview matrix since
we can assume the initial state is already the identity matrix. The
explicit initialization was only really necessary when running as a
clutter test because there the default matrix isn't the identity matrix.

Also some calls to cogl_orth to change the projection matrix have been
moved into the test entry point functions instead of calling this in the
paint function. Again the previous style probably came about because
with clutter we always had to re-assert the projection but now we are in
full control of the projection and can assume it doesn't need
re-asserting once set.

Acked-by: Luca Bruno <lucabru@src.gnome.org>