Rotate according to time not number of frames in the crate example
Previously the crate example incremented the angle of rotation of the
cube every frame so depending on the framerate the cube might rotate
too fast to see. This just changes it to calculate the rotation based
on the elapsed time using a GTimer. The rate that frames are drawn is
unaffected.

Reviewed-by: Robert Bragg <robert@linux.intel.com>